+++ openmsistream/data_file_io/data_file.py
@@ -1,5 +1,5 @@
 #imports
-from ..shared.logging import LogOwner
+from ..utilities.logging import LogOwner
 from ..utilities.misc import populated_kwargs
 
 class DataFile(LogOwner) :
