+++ openmsistream/data_file_io/data_file_download_directory.py
@@ -1,7 +1,7 @@
 #imports
 import datetime
 from kafkacrypto.message import KafkaCryptoMessage
-from ..shared.runnable import Runnable
+from ..running.runnable import Runnable
 from .config import DATA_FILE_HANDLING_CONST, RUN_OPT_CONST
 from .utilities import get_encrypted_message_key_and_value_filenames
 from .download_data_file import DownloadDataFileToDisk
