+++ openmsistream/data_file_io/data_file_upload_directory.py
@@ -3,10 +3,10 @@ import pathlib, datetime, time
 from threading import Lock
 from queue import Queue
 from ..utilities.misc import populated_kwargs
-from ..shared.config import UTIL_CONST
-from ..shared.my_thread import MyThread
-from ..shared.runnable import Runnable
-from ..shared.controlled_process import ControlledProcessSingleThread
+from ..utilities.config import UTIL_CONST
+from ..utilities.exception_tracking_thread import MyThread
+from ..running.runnable import Runnable
+from ..running.controlled_process import ControlledProcessSingleThread
 from ..my_kafka.producer_group import ProducerGroup
 from .config import RUN_OPT_CONST
 from .producer_file_registry import ProducerFileRegistry
