+++ openmsistream/data_file_io/upload_data_file.py
@@ -3,7 +3,7 @@ import traceback, time
 from threading import Thread
 from queue import Queue
 from hashlib import sha512
-from ..shared.runnable import Runnable
+from ..running.runnable import Runnable
 from ..my_kafka.producer_group import ProducerGroup
 from .config import RUN_OPT_CONST
 from .data_file_chunk import DataFileChunk
