+++ openmsistream/my_kafka/consumer_group.py
@@ -1,5 +1,5 @@
 #imports
-from ..shared.logging import LogOwner
+from ..utilities.logging import LogOwner
 from .my_consumer import MyConsumer
 
 class ConsumerGroup(LogOwner) :
