+++ openmsistream/osn/osn_stream_processor.py
@@ -1,4 +1,4 @@
-from ..shared.runnable import Runnable
+from ..running.runnable import Runnable
 from ..data_file_io.config import RUN_OPT_CONST
 from ..data_file_io.data_file_stream_processor import DataFileStreamProcessor
 from .config_file_parser import OSNConfigFileParser
