+++ test/rebuild_test_reference_data.py
@@ -1,6 +1,6 @@
 #imports
 import pathlib, logging, shutil, filecmp
-from openmsistream.shared.logging import Logger
+from openmsistream.utilities.logging import Logger
 from openmsistream.data_file_io.config import RUN_OPT_CONST
 from openmsistream.data_file_io.upload_data_file import UploadDataFile
 from openmsistream.my_kafka.serialization import DataFileChunkSerializer
