+++ test/unittests/test_data_file_chunk.py
@@ -1,7 +1,7 @@
 #imports
 import unittest, pathlib, logging
 from confluent_kafka.error import SerializationError
-from openmsistream.shared.logging import Logger
+from openmsistream.utilities.logging import Logger
 from openmsistream.data_file_io.config import RUN_OPT_CONST
 from openmsistream.data_file_io.upload_data_file import UploadDataFile
 from openmsistream.data_file_io.data_file_chunk import DataFileChunk
