+++ test/unittests/test_data_file_stream_processor.py
@@ -1,7 +1,7 @@
 #imports
 import unittest, time, pathlib, logging
-from openmsistream.shared.logging import Logger
-from openmsistream.shared.my_thread import MyThread
+from openmsistream.utilities.logging import Logger
+from openmsistream.utilities.exception_tracking_thread import MyThread
 from openmsistream.data_file_io.config import RUN_OPT_CONST
 from openmsistream.data_file_io.upload_data_file import UploadDataFile
 from openmsistream.data_file_io.data_file_stream_processor import DataFileStreamProcessor
