+++ test/unittests/test_download_data_file.py
@@ -1,7 +1,7 @@
 #imports
 import unittest, pathlib, logging, filecmp, shutil
 from hashlib import sha512
-from openmsistream.shared.logging import Logger
+from openmsistream.utilities.logging import Logger
 from openmsistream.data_file_io.config import RUN_OPT_CONST, DATA_FILE_HANDLING_CONST
 from openmsistream.data_file_io.data_file_chunk import DataFileChunk
 from openmsistream.data_file_io.upload_data_file import UploadDataFile
