+++ test/unittests/test_upload_data_file.py
@@ -1,7 +1,7 @@
 #imports
 import unittest, pathlib, logging
 from queue import Queue
-from openmsistream.shared.logging import Logger
+from openmsistream.utilities.logging import Logger
 from openmsistream.data_file_io.config import RUN_OPT_CONST
 from openmsistream.data_file_io.upload_data_file import UploadDataFile
 from config import TEST_CONST
